Moved dev tools to slider
